Fix permission flow: create session before SDK query to prevent race
Root cause: the SDK's Query constructor starts processing messages
(including canUseTool permission requests) immediately. Our code was
creating the session in activeSessions AFTER calling sdkQuery(), so
when canUseTool fired, activeSessions.get(sessionId) returned undefined
and the resolver was never stored — making Allow clicks silently fail.

Fix: move session creation before the sdkQuery() call so the session
is in the map when canUseTool is called.

Also:
- Re-add permissionMode to query options (CLI needs it to decide which
  tools require permission checks in the first place)
- Add E2E_REAL_SDK flag to test with real Claude SDK in E2E harness
- Add real-SDK integration test that verifies the full permission flow
  (prompt appears → Allow clicked → tool executes → no rejection)
